+++ internal/domain/repositories/bot_repo.go
@@ -7,6 +7,6 @@ import (
 )
 
 type BotRepository interface {
-	FindBotByTgIdAndType(ctx context.Context, tgid int32, botType entities.BotType) (*entities.BotWithChannel, error)
-	Create(ctx context.Context, tgid int32, botType entities.BotType) (*entities.BotWithChannel, error)
+	FindBotByTgIdAndType(ctx context.Context, tgid int64, botType entities.BotType) (*entities.BotWithChannel, error)
+	Create(ctx context.Context, tgid int64, botType entities.BotType) (*entities.BotWithChannel, error)
 }
